requirements.txt
python-telegram-bot 
requests
pillow
python-dotenv
pytz